Remove logStream metrics from LoggregatorClient
These metrics were originally added as part of the following story:

https://www.pivotaltracker.com/story/show/54779958

They were used to verify that log messages were not lost within the agent process. The story confirms that messages are never lost, thus the metrics are not needed anymore.

Plus, it was hacky to hang the logStream metrics off of LoggregatorClient.

Signed-off-by: Glenn Oppegard <goppegard@pivotallabs.com>