Ensure that TrafficController properly closes websocket connections
Previous behavior was that a CLI-TC websocket closure did not properly propogate, resulting in lack of TC-Doppler closures, resulting in leaked network connections, goroutines, and memory.

Complexity (additional goroutine, etc.) in WebsocketListener is a result of the fact that setting timing out network reads corrupts the connection, causing all subsequent reads to fail. This necessitated an externally-forced closure of the websocket connection during a blocking read, which then necessitated handling the error retrurned by the read in a special way.

[#89632832]