Fix metron/messageaggregator test
Test uncovered an earlier bug: if an events.Envelope has a nil EventType field, its GetEventType() method returns the lowest value for the enum.
Previously, this was a Heartbeat message, so MessageAggregator.Run dropped into its default case; now HTTPStart is lowest. Added an explicit nil check
to avoid this in the future.

Also updates BOSH specs to not reference sonde-go/control.

[#96706488]

Signed-off-by: John Tuley <jtuley@pivotal.io>