Read to detect syslog writer connection close
Writes to a connection that has been closed on the remote end can appear
to be successful even though the message never makes it across the wire.
When the syslog writer's connection gets into this state, it can fail to
reconnect in a timely manner.  Hanging out a read on the connection
helps to detect this condition more quickly.