Update locked_memory_limit
Since we're already doing string processing for 'unlimited', it made sense to use 'kernel' instead of -1 to
specify that we should use the kernel's default.

We also added a little validation to make sure that the value we get is either numeric or one of a known set
of valid string parameters.

[#115943629]

Signed-off-by: Sam Nelson <snelson@pivotal.io>