Set the status code according to the failure
The log access authorizer in CC returns 200, 401, 403, 404 and 500.
Reflect the actual status of the request except for 403 which should be
treated as a 404.

The main importance is to determine the difference between an actual 401 which
requires a refresh of the token or a 404 when the app is not
present/forbidden.