Configure syslog forwarder in the background
Restart rsyslog can take more than 30s in some cases. If that happens
monit will timeout starting metron_agent, when it could start properly.

For instance when using RELP, if the connection to the remote server is slow
of stale, or the remote RELP server is behind a load balancer with all
the backends down.