Use pprof as the endpoint signalling that a component is ready
Our integration test helpers now always wait for the pprof port to accept connections before
returning.  This means that the component being started will be ready by the time the helper
returns.

To facilitate this, we've had to move the pprof listening action to the end of each component's
main.

[#135625313]