Remove dea_logging_agent concerns from loggregator
The `logger` package is used only in the `dea_logging_agent`, which vendors
it. This commit removes `logger` from loggregator and instead pulls in an
updated `dea_logging_agent` submodule which includes it.

This commit also moves the integration tests for the dea agent into the
`dea_logging_agent` package.