Fix deadlock in truncating buffer
When the truncating buffer droppes messages it also emits some to notify downstream consumers that it has dropped messages. However, it tries to write to the channel and if that channel is full, it will block. It will not drop messages while this occurs, meaning anything writing can also get blocked.

[#140807909]