Consolidate calls to dropsonde/metrics behind shim
The dropsonde/metrics package is based on global state. As a result, it is
hard to test and conceals a type's true dependencies. This commit moves all
calls to metrics.SendValue into a shim, which is then injected into any
handler which requires it.

By moving the metrics behind a shim, we are now in a better position to
outright replace it with more recent and reliable code for metrics.

As part of the metrics work, this commit also groups all web socket concerns
into a type, which is injected into any handlers which require it. The web
socket type, WebSocketServer, likewise consumes the metrics shim to emit
metrics.