Move syslog daemon config
The syslog daemon config is not loggregator source code and as such has
no business in code.cloudfoundry.org/loggregator. This commit moves it
up as a top level entry in Bosh's src directory.

This commit also updates the configuration of Metron and Metron Windows
to ensure that both jobs consume a syslog daemon config package as part
of their configuration. Subsequent Metron package and job configuration
remains the same.