Add a drain script for the loggregator_trafficcontroller
We want to make sure the traffic controller is stopped first before any other services. We use Consul DNS for resolving the Cloud Controller addresses, but there is no guarantee that the Consul agent is not stopped before the traffic controller service which will cause errors.

When stopping the traffic controller VM we regularly see the following error message:

```
2018/01/23 15:13:28 Could not get app information: [Get https://cloud-controller-ng.service.cf.internal:9023/internal/v4/log_access/xxx: dial tcp: lookup cloud-controller-ng.service.cf.internal on 10.0.0.2:53: no such host
```