Remove grpc-throughputlb package
As far as we can tell this was brought into the code for a pretty minor
speed boost to our grpc connector (see [here](https://github.com/cloudfoundry/loggregator/pull/3) for more info).

The grpc balancer package is now deprecated, and updating this code to
work with the new grpc version looks very difficult. We do not think
this package is worth saving.

Signed-off-by: Carson Long <lcarson@vmware.com>