Refactor TrafficController Log Cache connection
`google.golang.org/grpc/balancer` is deprecated, so we cannot expect
to continue passing `grpc.WithBalancerName` in order to set our gRPC Log
Cache connection to use a round robin strategy.

gRPC does come with a round robin option builtin, and the new way to
activate it is with `grpc.WithDefaultServiceConfig`. See
https://github.com/grpc/grpc-go/blob/v1.48.0/examples/features/load_balancing/client/main.go.