Bump gRPC version
We want to bump our gRPC version to try out the module without the
grpc-throughputlb, which was the previous blocker on bumping the gRPC
version. However, bumping gRPC also bumps protobuf, which causes
failures due to namespace conflicts. Therefore a lot of dependencies
need to be bumped in order to get around the namespace conflict:

* Upgrade to go-loggregator v9; includes new protobuf generated code and
  switches to google protobuf library over (unsupported gogo/protobuf).
* Bump go-log-cache, sonde-go, noaa, and other libraries to support this
  change.

Also:

* Change the code in this module to use the google protobuf library for
  protobuf (de/)serialization.
* Rename grpc.proto to doppler.proto (more clear) and regenerate that
  code.
* Embed gRPC methods into implementations of gRPC servers to satisfy new
  requirements around future-proofing code. (MustEmbedUnimplemented...).