Remove grpc-throughputlb package (#453)
* Remove grpc-throughputlb package

As far as we can tell this was brought into the code for a pretty minor
speed boost to our grpc connector (see [here](https://github.com/cloudfoundry/loggregator/pull/3) for more info).

The grpc balancer package is now deprecated, and updating this code to
work with the new grpc version looks very difficult. We do not think
this package is worth saving.

Signed-off-by: Carson Long <lcarson@vmware.com>

* Update packages to remove grpc-throughputlb

Also remove `-mod=vendor` flag from many packaging scripts, as that's
the default for `go build`.

Signed-off-by: Matthew Kocher <mkocher@vmware.com>

* Refactor TrafficController Log Cache connection

`google.golang.org/grpc/balancer` is deprecated, so we cannot expect
to continue passing `grpc.WithBalancerName` in order to set our gRPC Log
Cache connection to use a round robin strategy.

gRPC does come with a round robin option builtin, and the new way to
activate it is with `grpc.WithDefaultServiceConfig`. See
https://github.com/grpc/grpc-go/blob/v1.48.0/examples/features/load_balancing/client/main.go.

* Bump gRPC version

We want to bump our gRPC version to try out the module without the
grpc-throughputlb, which was the previous blocker on bumping the gRPC
version. However, bumping gRPC also bumps protobuf, which causes
failures due to namespace conflicts. Therefore a lot of dependencies
need to be bumped in order to get around the namespace conflict:

* Upgrade to go-loggregator v9; includes new protobuf generated code and
  switches to google protobuf library over (unsupported gogo/protobuf).
* Bump go-log-cache, sonde-go, noaa, and other libraries to support this
  change.

Also:

* Change the code in this module to use the google protobuf library for
  protobuf (de/)serialization.
* Rename grpc.proto to doppler.proto (more clear) and regenerate that
  code.
* Embed gRPC methods into implementations of gRPC servers to satisfy new
  requirements around future-proofing code. (MustEmbedUnimplemented...).

Co-authored-by: Matthew Kocher <mkocher@vmware.com>