Fix race condition in the v1 ingestor server tests
There was a recurring race condition because of a goroutine that was
running the `Pusher` method on manager, and the BeforeEach which would
assign a new v1 `IngestorServer` to manager.

This change stops the goroutine after it's not needed, and waits for it
to close before completing the `It` test.