Fix flakey router test
Only start the spyAgent's gRPC server for tests that require it to be
started and/or wait for it to be started. This should just be the tests
related to metrics reporting, as that is all the router appears to
accept an agent address for.

Previously, we were starting the spyAgent's gRPC server in a separate
goroutine in the BeforeEach, which could cause flakes when the
individual tests completed too quickly, and caused the AfterEach block
to trigger and stop the gRPC server before the goroutine was scheduled.
The server attempting to start after being told to stop would throw an
error.

Signed-off-by: Matthew Kocher <mkocher@vmware.com>