Refactor src/trafficcontroller/internal/proxy/websocket_handler_test.go
The tests were failing consistently with the new gorilla/websocket
package. The nature of the failures indicated that the tests were at
fault rather than new functionality from the package.

This commit refactors the tests to use a websocket client in order to
connect to the test server in standard, supported way. Additionally, it
updates many of the tests that weren't actually testing what their
descriptions purported they were testing.

One test for "old-style keep-alives" was completely removed. The test
wasn't really testing for it, and we couldn't find any evidence that it
is a real thing that we should support.

wip